+++ src/config.php
@@ -95,6 +95,9 @@ return [
         // The amount of time in seconds to reserve a sendout job
         //'sendoutJobTtr' => 300,
 
+        // The priority to give the sendout cache job (the lower the number, the higher the priority)
+        //'sendoutJobPriority' => 10,
+
         // Enable GeoIP to geolocate contacts by their IP addresses
         //'geoIp' => false,